temp push

+++ includes/utils.h
@@ -3,9 +3,9 @@
 
 typedef struct s_vec3
 {
-	float	x;
-	float	y;
-	float	z;
+	double	x;
+	double	y;
+	double	z;
 }	t_vec3;
 
 typedef struct s_color
@@ -18,12 +18,12 @@ typedef struct s_color
 typedef struct s_ambient
 {
 	t_color	color;
-	float	lighting_ratio;
+	double	lighting_ratio;
 }	t_ambient;
 
 typedef struct s_camera
 {
-	float	fov;
+	double	fov;
 	t_vec3	pos;
 	t_vec3	normal;
 }	t_camera;
@@ -32,7 +32,7 @@ typedef struct s_light
 {
 	t_vec3	pos;
 	t_color	color;
-	float	brightness_ratio;
+	double	brightness_ratio;
 }	t_light;
 
 #endif
