+++ sources/math/vector_utils.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   vector_utils.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dmitry <dmitry@student.42.fr>              +#+  +:+       +#+        */
+/*   By: lorphan <lorphan@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/15 15:24:48 by lorphan           #+#    #+#             */
-/*   Updated: 2022/03/23 01:32:05 by dmitry           ###   ########.fr       */
+/*   Updated: 2022/03/24 12:37:56 by lorphan          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minirt.h"
 
-float	vec_length(t_vec *vec)
+double	vec_length(t_vec *vec)
 {
 	if (!vec)
 		return (0);
@@ -21,7 +21,7 @@ float	vec_length(t_vec *vec)
 
 t_vec	vec_normalize(t_vec *vec)
 {
-	float	temp;
+	double	temp;
 
 	temp = vec_length(vec);
 	vec->x /= temp;
